web-clipper/requirements.txt
trafilatura>=1.6,<2.0
requests>=2.31,<3.0
python-slugify>=8.0,<9.0
pyyaml>=6.0,<7.0
